requirements.txt
Flask>=0.11.1
Flask_JWT>=0.3.2
Flask_RESTful>=0.3.5
Flask_SQLAlchemy>=2.1
python_bcrypt>=0.3.1
